Improve retry and timeout for HTTP JSON endpoints
